requirements.txt
bpy
mathutils
pathlib
pyclothoid
scenariogeneration
pytest